Fix selecting a word in text box
